fix: Restoring a hidden window now restores the badge too
Fixes #215
